[FE][hotfix] 라우팅 경로 수정, 404 페이지 배경 적용 (#228)
* chore: 라우팅 경로 수정 안 된 부분 수정

* chore: NotFound 컴포넌트 배경 이미지 설정
